Se genera CRUD Usuarios

+++ src/service/recipe_service.py
@@ -11,7 +11,7 @@ mysql  = MySQL(app)
 def get_all_recipes():
     try:
         cursor = mysql.connection.cursor()
-        sql = 'SELECT id, name, short_description, long_description, ranking, type FROM tc_recipe'
+        sql = 'SELECT id, name, short_description, long_description, ranking, type_id FROM tc_recipe'
         cursor.execute(sql)
         data = cursor.fetchall()
         recipes = []
@@ -20,22 +20,25 @@ def get_all_recipes():
             recipes.append(recipe)
         return jsonify({'recetas':recipes, 'message':'Consulta exitosa', 'code':200})
     except Exception as ex:
+        print(ex)
         return jsonify({'recetas':{}, 'message':'Error al realizar la operación', 'code':400})
     
 """
-*** Consulta todas las recetas **
+*** Consulta una receta **
 """
 def get_one_recipe(id):
     try:
         cursor = mysql.connection.cursor()
-        sql = "SELECT id, name, short_description, long_description, ranking, type FROM tc_recipe WHERE id = '{0}'".format(id)
+        sql = "SELECT id, name, short_description, long_description, ranking, type_id FROM tc_recipe WHERE id = '{0}'".format(id)
         cursor.execute(sql)
         data = cursor.fetchall()
+        print(data)
         if len(data) == 0:
             return jsonify({'recetas':{}, 'message':'Receta no encontrada', 'code':404})
         recipe = {'id':data[0][0],'name':data[0][1],'short_description':data[0][2],'long_description':data[0][3],'ranking':data[0][4],'type':data[0][5]}
         return jsonify({'recetas':recipe, 'message':'Consulta exitosa', 'code':200})
     except Exception as ex:
+        print(ex)
         return jsonify({'recetas':{}, 'message':'Error al realizar la operación', 'code':400})
 
 """
@@ -45,12 +48,18 @@ def update_recipe(data):
     try:
         # Logica de Operación
         cursor = mysql.connection.cursor()
-        sql = "UPDATE tc_recipe SET name = %s, short_description = %s, long_description = %s, ranking = %s, type = %s WHERE id = %s"
-        values = (data['name'], data['short_description'], data['long_description'], data['ranking'], data['type'], data['id'])
-        cursor.execute(sql,values)
-        mysql.connection.commit()
-        cursor.fetchall()
-        return jsonify({'receta':{}, 'message':'Actualización exitosa', 'code':200})
+        sqlGet = "SELECT id FROM tc_recipe WHERE id = '{0}'".format(data['id'])
+        cursor.execute(sqlGet)
+        res = cursor.fetchall()
+        if len(res) == 0:
+            return jsonify({'receta':{}, 'message':'Receta no encontrada', 'code':404})
+        else : 
+            sql = "UPDATE tc_recipe SET name = %s, short_description = %s, long_description = %s, ranking = %s, type_id = %s WHERE id = %s"
+            values = (data['name'], data['short_description'], data['long_description'], data['ranking'], data['type'], data['id'])
+            cursor.execute(sql,values)
+            mysql.connection.commit()
+            cursor.fetchall()
+            return jsonify({'receta':{}, 'message':'Actualización exitosa', 'code':200})
     except Exception as ex:
         print(ex)
         return jsonify({'receta':{}, 'message':'Error al realizar la operación', 'code':400})
@@ -62,11 +71,17 @@ def delete_recipe(id):
     try:
         # Logica de Operación
         cursor = mysql.connection.cursor()
-        sql = "DELETE FROM tc_recipe WHERE id = {0}".format(id)
-        cursor.execute(sql)
-        mysql.connection.commit()
-        cursor.fetchall()
-        return jsonify({'receta':{}, 'message':'Borrado exitoso', 'code':200})
+        sqlGet = "SELECT id FROM tc_recipe WHERE id = '{0}'".format(id)
+        cursor.execute(sqlGet)
+        data = cursor.fetchall()
+        if len(data) == 0:
+            return jsonify({'receta':{}, 'message':'Receta no encontrada', 'code':404})
+        else: 
+            sql = "DELETE FROM tc_recipe WHERE id = {0}".format(id)
+            cursor.execute(sql)
+            mysql.connection.commit()
+            cursor.fetchall()
+            return jsonify({'receta':{}, 'message':'Borrado exitoso', 'code':200})
     except Exception as ex:
         print(ex)
         return jsonify({'receta':{}, 'message':'Error al realizar la operación', 'code':400})
@@ -78,7 +93,7 @@ def insert_recipe(data):
     try:
         # Logica de Operación
         cursor = mysql.connection.cursor()
-        sql = "INSERT INTO tc_recipe (name, short_description, long_description, ranking, type) VALUES (%s, %s, %s, %s, %s)"
+        sql = "INSERT INTO tc_recipe (name, short_description, long_description, ranking, type_id) VALUES (%s, %s, %s, %s, %s)"
         values = (data['name'], data['short_description'], data['long_description'], data['ranking'], data['type'])
         cursor.execute(sql, values)
         mysql.connection.commit()
